feat(tree shaking): Start on implementing code splitting
