+++ flags/allFlags.js
@@ -1,0 +1,258 @@
+export { default as AD } from './AD'
+export { default as AE } from './AE'
+export { default as AFRUN } from './AFRUN'
+export { default as AG } from './AG'
+export { default as AI } from './AI'
+export { default as AL } from './AL'
+export { default as AM } from './AM'
+export { default as AMS } from './AMS'
+export { default as AO } from './AO'
+export { default as AQ } from './AQ'
+export { default as AR } from './AR'
+export { default as AS } from './AS'
+export { default as AT } from './AT'
+export { default as AU } from './AU'
+export { default as AW } from './AW'
+export { default as AX } from './AX'
+export { default as AZ } from './AZ'
+export { default as BA } from './BA'
+export { default as BB } from './BB'
+export { default as BD } from './BD'
+export { default as BE } from './BE'
+export { default as BF } from './BF'
+export { default as BG } from './BG'
+export { default as BH } from './BH'
+export { default as BI } from './BI'
+export { default as BJ } from './BJ'
+export { default as BL } from './BL'
+export { default as BM } from './BM'
+export { default as BN } from './BN'
+export { default as BO } from './BO'
+export { default as BQ_BO } from './BQ-BO'
+export { default as BQ_SA } from './BQ-SA'
+export { default as BQ_SE } from './BQ-SE'
+export { default as BR } from './BR'
+export { default as BS } from './BS'
+export { default as BT } from './BT'
+export { default as BW } from './BW'
+export { default as BY } from './BY'
+export { default as BZ } from './BZ'
+export { default as CA } from './CA'
+export { default as CC } from './CC'
+export { default as CD } from './CD'
+export { default as CF } from './CF'
+export { default as CG } from './CG'
+export { default as CH } from './CH'
+export { default as CI } from './CI'
+export { default as CK } from './CK'
+export { default as CL } from './CL'
+export { default as CM } from './CM'
+export { default as CN } from './CN'
+export { default as CO } from './CO'
+export { default as CR } from './CR'
+export { default as CU } from './CU'
+export { default as CV } from './CV'
+export { default as CW } from './CW'
+export { default as CX } from './CX'
+export { default as CY } from './CY'
+export { default as CZ } from './CZ'
+export { default as DE } from './DE'
+export { default as DJ } from './DJ'
+export { default as DK } from './DK'
+export { default as DM } from './DM'
+export { default as DO } from './DO'
+export { default as EC } from './EC'
+export { default as EE } from './EE'
+export { default as EG } from './EG'
+export { default as EH } from './EH'
+export { default as ER } from './ER'
+export { default as ES } from './ES'
+export { default as ET } from './ET'
+export { default as EU } from './EU'
+export { default as FI } from './FI'
+export { default as FJ } from './FJ'
+export { default as FK } from './FK'
+export { default as FM } from './FM'
+export { default as FO } from './FO'
+export { default as FR } from './FR'
+export { default as GA } from './GA'
+export { default as GB_ENG } from './GB-ENG'
+export { default as GB_SCT } from './GB-SCT'
+export { default as GB_UKM } from './GB-UKM'
+export { default as GB_WLS } from './GB-WLS'
+export { default as GB_NIR } from './GB-NIR'
+export { default as GD } from './GD'
+export { default as GE } from './GE'
+export { default as GF } from './GF'
+export { default as GG } from './GG'
+export { default as GH } from './GH'
+export { default as GI } from './GI'
+export { default as GL } from './GL'
+export { default as GM } from './GM'
+export { default as GN } from './GN'
+export { default as GP } from './GP'
+export { default as GQ } from './GQ'
+export { default as GR } from './GR'
+export { default as GS } from './GS'
+export { default as GT } from './GT'
+export { default as GU } from './GU'
+export { default as GW } from './GW'
+export { default as GY } from './GY'
+export { default as HK } from './HK'
+export { default as HM } from './HM'
+export { default as HN } from './HN'
+export { default as HR } from './HR'
+export { default as HT } from './HT'
+export { default as HU } from './HU'
+export { default as ID } from './ID'
+export { default as IE } from './IE'
+export { default as IL } from './IL'
+export { default as IM } from './IM'
+export { default as IN } from './IN'
+export { default as IO } from './IO'
+export { default as IQ } from './IQ'
+export { default as IR } from './IR'
+export { default as IS } from './IS'
+export { default as IT } from './IT'
+export { default as JE } from './JE'
+export { default as JM } from './JM'
+export { default as JO } from './JO'
+export { default as JP } from './JP'
+export { default as KE } from './KE'
+export { default as KG } from './KG'
+export { default as KH } from './KH'
+export { default as KI } from './KI'
+export { default as KM } from './KM'
+export { default as KN_SK } from './KN-SK'
+export { default as KP } from './KP'
+export { default as KR } from './KR'
+export { default as KW } from './KW'
+export { default as KY } from './KY'
+export { default as KZ } from './KZ'
+export { default as LA } from './LA'
+export { default as LB } from './LB'
+export { default as LC } from './LC'
+export { default as LI } from './LI'
+export { default as LK } from './LK'
+export { default as LR } from './LR'
+export { default as LS } from './LS'
+export { default as LT } from './LT'
+export { default as LU } from './LU'
+export { default as LV } from './LV'
+export { default as LY } from './LY'
+export { default as MA } from './MA'
+export { default as MC } from './MC'
+export { default as MD } from './MD'
+export { default as ME } from './ME'
+export { default as MF } from './MF'
+export { default as MG } from './MG'
+export { default as MH } from './MH'
+export { default as MK } from './MK'
+export { default as ML } from './ML'
+export { default as MM } from './MM'
+export { default as MN } from './MN'
+export { default as MO } from './MO'
+export { default as MP } from './MP'
+export { default as MQ } from './MQ'
+export { default as MR } from './MR'
+export { default as MS } from './MS'
+export { default as MT } from './MT'
+export { default as MU } from './MU'
+export { default as MV } from './MV'
+export { default as MW } from './MW'
+export { default as MX } from './MX'
+export { default as MY } from './MY'
+export { default as MZ } from './MZ'
+export { default as NA } from './NA'
+export { default as NC } from './NC'
+export { default as NE } from './NE'
+export { default as NF } from './NF'
+export { default as NG } from './NG'
+export { default as NI } from './NI'
+export { default as NL } from './NL'
+export { default as NO } from './NO'
+export { default as NP } from './NP'
+export { default as NR } from './NR'
+export { default as NU } from './NU'
+export { default as NZ } from './NZ'
+export { default as OM } from './OM'
+export { default as PA } from './PA'
+export { default as PE } from './PE'
+export { default as PF } from './PF'
+export { default as PG } from './PG'
+export { default as PH } from './PH'
+export { default as PK } from './PK'
+export { default as PL } from './PL'
+export { default as PM } from './PM'
+export { default as PN } from './PN'
+export { default as PR } from './PR'
+export { default as PS } from './PS'
+export { default as PT } from './PT'
+export { default as PW } from './PW'
+export { default as PY } from './PY'
+export { default as QA } from './QA'
+export { default as RAINBOW } from './RAINBOW'
+export { default as RE } from './RE'
+export { default as RO } from './RO'
+export { default as RS } from './RS'
+export { default as RU } from './RU'
+export { default as RW } from './RW'
+export { default as SA } from './SA'
+export { default as SB } from './SB'
+export { default as SC } from './SC'
+export { default as SD } from './SD'
+export { default as SE } from './SE'
+export { default as SG } from './SG'
+export { default as SH } from './SH'
+export { default as SI } from './SI'
+export { default as SJ } from './SJ'
+export { default as SK } from './SK'
+export { default as SL } from './SL'
+export { default as SM } from './SM'
+export { default as SN } from './SN'
+export { default as SO } from './SO'
+export { default as SR } from './SR'
+export { default as SS } from './SS'
+export { default as ST } from './ST'
+export { default as SV } from './SV'
+export { default as SX } from './SX'
+export { default as SY } from './SY'
+export { default as SZ } from './SZ'
+export { default as TC } from './TC'
+export { default as TD } from './TD'
+export { default as TF } from './TF'
+export { default as TG } from './TG'
+export { default as TH } from './TH'
+export { default as TJ } from './TJ'
+export { default as TK } from './TK'
+export { default as TL } from './TL'
+export { default as TM } from './TM'
+export { default as TN } from './TN'
+export { default as TO } from './TO'
+export { default as TR } from './TR'
+export { default as TT } from './TT'
+export { default as TV } from './TV'
+export { default as TW } from './TW'
+export { default as TZ } from './TZ'
+export { default as UA } from './UA'
+export { default as UG } from './UG'
+export { default as UM } from './UM'
+export { default as UNASUR } from './UNASUR'
+export { default as US } from './US'
+export { default as UY } from './UY'
+export { default as UZ } from './UZ'
+export { default as VA } from './VA'
+export { default as VC } from './VC'
+export { default as VE } from './VE'
+export { default as VG } from './VG'
+export { default as VI } from './VI'
+export { default as VN } from './VN'
+export { default as VU } from './VU'
+export { default as WF } from './WF'
+export { default as WS } from './WS'
+export { default as YE } from './YE'
+export { default as ZA } from './ZA'
+export { default as ZM } from './ZM'
+export { default as ZW } from './ZW'
+export { default as AF } from './AF'
+export { default as BV } from './BV'
