+++ index.js
@@ -1,18 +1,7 @@
-import codeList from './countryCodeList.json'
-import flags from './flags'
+import { flagList } from './flagList'
+import * as flags from './flags/allFlags'
 
-function isoToCountryCode(isoCode, keyToGet = 'alpha2') {
-  if (isoCode !== undefined) {
-    const alpha2Code = codeList.find((countryObj) => (
-      countryObj.alpha2 === isoCode
-      || countryObj.alpha3 === isoCode
-      || countryObj.numeric === isoCode
-    ))[keyToGet]
-    return alpha2Code
-  }
-
-  throw new Error('This code returns undefined, see https://www.flagpack.xyz/docs/flag-index for all the available codes.')
-}
+export { isoToCountryCode } from './isoToCountryCode'
 
 function imageUrl(assetCode, size) {
   const formatAssetCode = assetCode.replace('-', '_')
@@ -24,8 +13,5 @@ function imageUrl(assetCode, size) {
 }
 
 export {
-  codeList,
-  isoToCountryCode,
-  imageUrl,
-  flags
+  imageUrl
 }
