+++ package.json
@@ -2,6 +2,7 @@
   "name": "flagpack-core",
   "description": "",
   "version": "1.0.0",
+  "sideEffects": false,
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Yummygum/flagpack-core.git"
@@ -20,7 +21,11 @@
     "url": "https://github.com/Yummygum/flag-pack-core/issues"
   },
   "homepage": "https://github.com/yummygum",
-  "main": "dist/bundle.js",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "files": [
+    "dist"
+  ],
   "scripts": {
     "build": " NODE_ENV=production rollup -c",
     "dev": "node generateCodeList.js && node generateFlags.js && NODE_ENV=development rollup -cw"
