+++ rollup.config.js
@@ -4,6 +4,7 @@ import commonjs from 'rollup-plugin-commonjs'
 import resolve from 'rollup-plugin-node-resolve'
 import smartAsset from 'rollup-plugin-smart-asset'
 import minify from 'rollup-plugin-babel-minify'
+import pkg from './package.json'
 
 const dev = 'development'
 const prod = 'production'
@@ -20,20 +21,34 @@ const plugins = [
     url: 'copy',
     assetsPath: 'flags',
     useHash: true,
-    keepImport: true
+    keepImport: true,
   })
 ]
 
 if (env === prod) {
-  plugins.push(minify())
+  // plugins.push(minify())
 }
 
 export default {
   plugins,
-  input: 'index.js',
-  output: {
-    file: 'dist/bundle.js',
-    format: 'cjs',
-    name: 'flagpack-core'
-  }
+  input: [
+    'index.js',
+    'flags/allFlags.js'
+  ],
+  output: [
+    {
+      // file: pkg.main,
+      format: 'cjs',
+      name: 'flagpack-core',
+      dir: 'dist',
+      entryFileNames: '[name].[format].js'
+    },
+    {
+      // file: pkg.module,
+      format: 'es',
+      name: 'flagpack-core',
+      dir: 'dist',
+      entryFileNames: '[name].esm.js'
+    },
+  ]
 }
